feat: src directory
