+++ src/directors/continuous_director.py
@@ -1,9 +1,9 @@
 import time
 
-from config import load_config
-from directors.base_director import BaseDirector
-from publisher import Publisher
-from utils import calculate_acceptable_box, calculate_center_bbox
+from src.config import load_config
+from src.directors.base_director import BaseDirector
+from src.publisher import Publisher
+from src.utils import calculate_acceptable_box, calculate_center_bbox
 
 
 class ContinuousDirector(BaseDirector):
@@ -14,7 +14,9 @@ class ContinuousDirector(BaseDirector):
     movement_detection_start_time = None
 
     # This method is called to process each frame
-    def process_frame(self, hostname: str, bounding_box: list, frame_shape, publisher: Publisher):
+    def process_frame(
+        self, hostname: str, bounding_box: list, frame_shape, publisher: Publisher
+    ):
         """
         Based on received bounding box, this method tells the arm where to move the keep the subject in the acceptable box..
         It does this by continuously sending polar pan start and a direction until the subject is in the acceptable box.
@@ -61,10 +63,7 @@ class ContinuousDirector(BaseDirector):
                 self.movement_detection_start_time = current_time
 
             # Check if they've been outside for at least the confirmation delay
-            if (
-                current_time - self.movement_detection_start_time
-                < confirmation_delay
-            ):
+            if current_time - self.movement_detection_start_time < confirmation_delay:
                 return
             change_in_x = 0
             change_in_y = 0
