+++ src/utils.py
@@ -37,7 +37,7 @@ def calculate_acceptable_box(
     - frame_height
     """
     if acceptable_box_percent is None:
-        from config import DEFAULT_CONFIG
+        from src.config import DEFAULT_CONFIG
 
         acceptable_box_percent = (
             acceptable_box_percent or DEFAULT_CONFIG["acceptable_box_percent"]
